move helper methods to Transforms
